Use ops diffraction kernel
Replace Diffraction PSF 3D with ops kernelDiffraction.  I am not
familiar with the parameterization of 'Diffraction PSF 3D' (that code
was written by Daniel White) so I did not attempt to replicate the old
approach exactly.  Instead I chose parameters for kernelDiffraction such
that the resulting PSF is similar and the chirp img is attenuated by roughly
the same amopunt after convolution, and can be used to demonstrate restoration
with the various flavors of Richardson Lucy deconvolution.

Note: ops KernelDiffraction does not normalize the sum of the PSF to 1,
thus that step had to be implemented explicitly.