Create outputs before calling deconvolution
The current version of the Richardson Lucy ops require the output to be
created before calling the function and then used as an input to the
functions.  This approach isn't ideal and occurred when trying to implement these
ops as computers.  What we really want is Hybrids, that is the ops
should be callable as both Functions and Computers.

The next generation version of ops will handle the computer and function
flavors of ops more ellegantly.  When the new version of ops is ready I
can take another pass at the deconvolution script examples and change back
to the simpler function interface.